# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Keep this title stable: generated package paths depend on it.
  title: Api
  version: 1.0.0
  description: School administration API
servers:
  - url: /api
    description: Same-origin API
...
paths:
...
  /classes:
    get:
      operationId: listClasses
      tags: [classes]
      summary: List school classes
      responses:
        "200":
          description: Class list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SchoolClass"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createClass
      tags: [classes]
      summary: Create a class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassInput"
      responses:
        "201":
          description: Created class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchoolClass"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /classes/{id}:
    patch:
      operationId: updateClass
      tags: [classes]
      summary: Update a class
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassInput"
      responses:
        "200":
          description: Updated class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchoolClass"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      operationId: deleteClass
      tags: [classes]
      summary: Delete a class with no enrolled students
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Class deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: Search and filter the student register
      parameters:
        - name: academicYear
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [name, code, newest]
        - name: search
          in: query
          schema:
            type: string
        - name: classId
          in: query
          schema:
            type: integer
            minimum: 1
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Student list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createStudent
      tags: [students]
      summary: Enroll a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Created student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /students/register:
    get:
      operationId: getStudentRegister
      tags: [students]
      summary: Search the student register with server-side pagination
      parameters:
        - name: academicYear
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [name, code, newest]
        - name: search
          in: query
          schema:
            type: string
        - name: classId
          in: query
          schema:
            type: integer
            minimum: 1
        - name: active
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
      responses:
        "200":
          description: Paginated student records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentRegister"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /students/{id}:
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update student details
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      operationId: archiveStudent
      tags: [students]
      summary: Archive a student without deleting historical records
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Student archived
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /barcode/{code}:
    get:
      operationId: findStudentByBarcode
      tags: [students]
      summary: Find an enrolled student by barcode or student number
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 100
      responses:
        "200":
          description: Matching student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /subjects:
    get:
      operationId: listSubjects
      tags: [subjects]
      summary: List subjects, optionally for one class
      parameters:
        - name: classId
          in: query
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Subject list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createSubject
      tags: [subjects]
      summary: Add a subject to a class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectInput"
      responses:
        "201":
          description: Created subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /subjects/{id}:
    patch:
      operationId: updateSubject
      tags: [subjects]
      summary: Update a subject
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectInput"
      responses:
        "200":
          description: Updated subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      operationId: deleteSubject
      tags: [subjects]
      summary: Delete a subject with no exam history
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Subject deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
  responses:
    BadRequest:
      description: The submitted data is invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Unauthorized:
      description: Sign in is required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Forbidden:
      description: This account does not have permission
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: The requested record does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
...
    ApiError:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    SchoolClass:
      type: object
      properties:
        id:
          type: integer
        gradeLevel:
          type: string
        section:
          type: string
        academicYear:
          type: string
        room:
          type: [string, "null"]
        active:
          type: boolean
        studentCount:
          type: integer
      required: [id, gradeLevel, section, academicYear, room, active, studentCount]
    ClassInput:
      type: object
      properties:
        gradeLevel:
          type: string
          minLength: 1
          maxLength: 100
        section:
          type: string
          minLength: 1
          maxLength: 30
        academicYear:
          type: string
          minLength: 4
          maxLength: 30
        room:
          type: [string, "null"]
          maxLength: 100
        active:
          type: boolean
      required: [gradeLevel, section, academicYear, active]
    Student:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        studentCode:
          type: string
        barcode:
          type: string
        gender:
          type: [string, "null"]
        dateOfBirth:
          type: [string, "null"]
          format: date
        guardianName:
          type: [string, "null"]
        guardianPhone:
          type: [string, "null"]
        address:
          type: [string, "null"]
        classId:
          type: integer
        className:
          type: string
        section:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        [id, fullName, studentCode, barcode, gender, dateOfBirth, guardianName, guardianPhone, address, classId, className, section, active, createdAt]
    StudentInput:
      type: object
      properties:
        fullName:
          type: string
          minLength: 2
          maxLength: 160
        studentCode:
          type: string
          minLength: 1
          maxLength: 60
        gender:
          type: [string, "null"]
          maxLength: 30
        dateOfBirth:
          type: [string, "null"]
          format: date
        guardianName:
          type: [string, "null"]
          maxLength: 160
        guardianPhone:
          type: [string, "null"]
          maxLength: 50
        address:
          type: [string, "null"]
          maxLength: 500
        classId:
          type: integer
          minimum: 1
        active:
          type: boolean
      required: [fullName, studentCode, classId, active]
    StudentRegister:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: "#/components/schemas/Student"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
      required: [records, total, page, pageSize]
    Subject:
      type: object
      properties:
        id:
          type: integer
        classId:
          type: integer
        className:
          type: string
        section:
          type: string
        name:
          type: string
        code:
          type: [string, "null"]
        totalMarks:
          type: number
        passingMarks:
          type: number
      required: [id, classId, className, section, name, code, totalMarks, passingMarks]
    SubjectInput:
      type: object
      properties:
        classId:
          type: integer
          minimum: 1
        name:
          type: string
          minLength: 1
          maxLength: 120
        code:
          type: [string, "null"]
          maxLength: 40
        totalMarks:
          type: number
          minimum: 1
        passingMarks:
          type: number
          minimum: 0
      required: [classId, name, totalMarks, passingMarks]
...