# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * School administration API
 * OpenAPI spec version: 1.0.0
 */

export interface Student {
  id: number;
  fullName: string;
  studentCode: string;
  barcode: string;
  /** @nullable */
  gender: string | null;
  /** @nullable */
  dateOfBirth: Date | null;
  /** @nullable */
  guardianName: string | null;
  /** @nullable */
  guardianPhone: string | null;
  /** @nullable */
  address: string | null;
  classId: number;
  className: string;
  section: string;
  active: boolean;
  createdAt: Date;
}