# lib/api-zod/src/generated/types/studentInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * School administration API
 * OpenAPI spec version: 1.0.0
 */

export interface StudentInput {
  /**
     * @minLength 2
     * @maxLength 160
     */
  fullName: string;
  /**
     * @minLength 1
     * @maxLength 60
     */
  studentCode: string;
  /**
     * @maxLength 30
     * @nullable
     */
  gender?: string | null;
  /** @nullable */
  dateOfBirth?: Date | null;
  /**
     * @maxLength 160
     * @nullable
     */
  guardianName?: string | null;
  /**
     * @maxLength 50
     * @nullable
     */
  guardianPhone?: string | null;
  /**
     * @maxLength 500
     * @nullable
     */
  address?: string | null;
  /** @minimum 1 */
  classId: number;
  active: boolean;
}